--- Ontario-Crime-Mapper-Frontend/ocmf/Components/Login_Form/LoginForm.module.css
@@ -1,144 +1,0 @@
-.loginContainer {
-  display: flex;
-  height: 100%;
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-  background-color: rgb(79, 76, 76);
-}
-.loginDiv {
-  background-color: black;
-  /* padding: 20px; */
-   display: flex;
-  /*justify-content: center;
-  align-items: center; */
-  width: 60%;
-  height: 70%;
-  gap: 10px;
-  border-radius: 10px;
-
-}
-
-.loginDiv > * {
-    width: 50%;
-}
-
-.imageDiv {
-    display: flex;
-    justify-content: center;
-    /*height: 50%; */
-}
-
-.loginImage {
-    width: 100%;
-    height: 100%;
-    border-radius: 10px;
-}
-
-
-.formDiv {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    /* padding: 50px 50px; */
-}
-
-.formDiv > * {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  gap: 10px;
-  margin: 0;
-
-}
-
-.formDiv > h2 {
-    text-align: center;
-    padding-bottom: 10px;
-}
-
-.form {
-    padding: 30px 80px 50px 80px;
-    
-}
-
-.form > input {
-    height: 25px;
-    min-height: 25px;
-    min-width: 250px;
-} 
-
-.form > button {
-    padding: 10px;
-    font-size: 16px;
-    font-weight: bold;
-}
-.form > button:hover {
-    cursor: pointer;
-}
-
-.headerDiv {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: row;
-}
-
-.headerDiv > h2 {
-    width: 80%;
-    text-align: center;
-    padding-right: 50px;
-}
-
-.headerDiv div:first-child {
-    display: flex;
-    justify-content: flex-start;
-    width: 10%;
-    padding: 0 10px;
-}
-
-.loginPageLinkDiv {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-self: center;
-    font-size: 14px;
-    gap: 4px;
-  }
-  
-  .loginPageLinkDiv > a {
-    text-decoration: underline;
-  }
-
-.errmsg {
-    background-color: lightpink;
-    color: firebrick;
-    font-weight: bold;
-    padding: 0.5rem;
-    margin-bottom: 0.5rem;
-    text-align: center;
-}
-
-.offScreen {
-    display: none;
-}
-
-@media screen and (max-width: 1500px) {
-    .loginDiv {
-        width: 70vw;
-        min-width: 400px;
-        max-width: 600px;
-        /* max-width: ; */
-    }
-    .imageDiv {
-        display: none;
-        width: 0;
-    }
-    .formDiv {
-        width: 100%;
-    }
-    .form {
-        padding: 20px;
-    }
-}
